+++ shared/src/config.rs
@@ -1,12 +1,12 @@
 use {
     crate::five8::{pubkey_decode, signature_decode},
-    base64::{engine::general_purpose::STANDARD as base64_engine, Engine},
+    base64::{Engine, engine::general_purpose::STANDARD as base64_engine},
     human_size::Size,
     regex::Regex,
     rustls::pki_types::{CertificateDer, PrivateKeyDer, PrivatePkcs8KeyDer},
     serde::{
-        de::{self, Deserializer},
         Deserialize,
+        de::{self, Deserializer},
     },
     solana_sdk::{pubkey::Pubkey, signature::Signature},
     std::{